requirements.txt.txt
fastapi
uvicorn
line-bot-sdk==3.7.0
